+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/spec/builtin/msg/req/BuiltinMsg0100V2013Alias.java
@@ -2,10 +2,8 @@ package io.github.hylexus.jt.jt808.spec.builtin.msg.req;
 
 import io.github.hylexus.jt.annotation.BuiltinComponent;
 import io.github.hylexus.jt.jt808.support.annotation.msg.req.Jt808RequestBody;
-import io.github.hylexus.jt.jt808.support.annotation.msg.req.RequestField;
+import io.github.hylexus.jt.jt808.support.annotation.msg.req.RequestFieldAlias;
 import lombok.Data;
-
-import static io.github.hylexus.jt.jt808.support.data.MsgDataType.*;
 
 /**
  * 内置终端注册消息体(2013)
@@ -15,32 +13,32 @@ import static io.github.hylexus.jt.jt808.support.data.MsgDataType.*;
 @Data
 @Jt808RequestBody
 @BuiltinComponent
-public class BuiltinMsg0100V2013 {
+public class BuiltinMsg0100V2013Alias {
     // 1. [0-2) WORD 省域ID
-    @RequestField(order = 1, startIndex = 0, dataType = WORD)
+    @RequestFieldAlias.Word(order = 1)
     private int provinceId;
 
-    // 2. [2-4) WORD 省域ID
-    @RequestField(order = 2, startIndex = 2, dataType = WORD)
+    // 2. [2-4) WORD 市县域ID
+    @RequestFieldAlias.Word(order = 2)
     private int cityId;
 
     // 3. [4-9) BYTE[5] 制造商ID
-    @RequestField(order = 3, startIndex = 4, dataType = BYTES, length = 5)
+    @RequestFieldAlias.Bytes(order = 3, length = 5)
     private String manufacturerId;
 
     // 4. [9-29) BYTE[20] 终端型号
-    @RequestField(order = 4, startIndex = 9, dataType = BYTES, length = 20)
+    @RequestFieldAlias.Bytes(order = 4, length = 20)
     private String terminalType;
 
     // 5. [29-36) BYTE[7] 终端ID
-    @RequestField(order = 5, startIndex = 17, dataType = BYTES, length = 7)
+    @RequestFieldAlias.Bytes(order = 5, length = 7)
     private String terminalId;
 
     // 6. [36]   BYTE    车牌颜色
-    @RequestField(order = 6, startIndex = 36, dataType = BYTE)
+    @RequestFieldAlias.Byte(order = 6)
     private byte color;
 
     // 7. [37,n)   String    车辆标识
-    @RequestField(order = 7, startIndex = 37, dataType = STRING, lengthExpression = "#ctx.msgBodyLength() - 37")
+    @RequestFieldAlias.String(order = 7, lengthExpression = "#ctx.msgBodyLength() - 37")
     private String carIdentifier;
 }