+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/annotation/msg/req/RequestField.java
@@ -68,4 +68,5 @@ public @interface RequestField {
 
     Class<? extends Jt808FieldDeserializer<?>> customerFieldDeserializerClass() default Jt808FieldDeserializer.PlaceholderFieldDeserializer.class;
 
+    String desc() default "";
 }