+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/data/deserialize/DefaultJt808FieldDeserializerRegistry.java
@@ -47,6 +47,8 @@ public class DefaultJt808FieldDeserializerRegistry implements Jt808FieldDeserial
 
         registry.registerConverter(new ByteBufContainerFieldDeserializer());
 
+        registry.registerConverter(new BitOperatorFieldDeserializer());
+
     }
 
     @Override