+++ jt-808-server-support/src/main/java/io/github/hylexus/jt/jt808/support/data/deserialize/impl/StringFieldDeserializer.java
@@ -17,7 +17,8 @@ import java.util.Set;
  */
 public class StringFieldDeserializer implements Jt808FieldDeserializer<String> {
     private static final Set<RequestMsgConvertibleMetadata> CONVERTIBLE_METADATA_SET = Set.of(
-            ConvertibleMetadata.forJt808RequestMsgDataType(MsgDataType.STRING, String.class)
+            ConvertibleMetadata.forJt808RequestMsgDataType(MsgDataType.STRING, String.class),
+            ConvertibleMetadata.forJt808RequestMsgDataType(MsgDataType.BYTES, String.class)
     );
 
     private final Charset charset;
@@ -37,7 +38,7 @@ public class StringFieldDeserializer implements Jt808FieldDeserializer<String> {
 
     @Override
     public String deserialize(ByteBuf byteBuf, MsgDataType msgDataType, int start, int length) {
-        if (msgDataType == MsgDataType.STRING) {
+        if (msgDataType == MsgDataType.STRING || msgDataType == MsgDataType.BYTES) {
             return JtProtocolUtils.readString(byteBuf, length, charset);
         }
         throw new Jt808AnnotationArgumentResolveException("Cannot convert DataType from " + msgDataType + " to String");